评论部分

+++ user_backend/src/main/java/com/albumsystem/user_backend/pojo/CommentReport.java
@@ -18,7 +18,7 @@ public class CommentReport {
     private Integer commentId;
     private Integer userId;
     private String reason;
-    private Integer status;
+    private Short status;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Shanghai")
     private Date time;
 }
